+++ sdk/keyvault/azure-security-keyvault-jca/src/main/java/com/azure/security/keyvault/jca/AuthClient.java
@@ -1,14 +1,14 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-
 package com.azure.security.keyvault.jca;
 
-import com.azure.security.keyvault.jca.rest.OAuthToken;
+import com.azure.security.keyvault.jca.model.OAuthToken;
 
 import java.util.HashMap;
 import java.util.logging.Logger;
 
 import static java.util.logging.Level.FINER;
+import static java.util.logging.Level.INFO;
 
 /**
  * The REST client specific to getting an access token for Azure REST APIs.
@@ -49,7 +49,7 @@ class AuthClient extends DelegateRestClient {
      * Stores the OAuth2 managed identity URL.
      */
     private static final String OAUTH2_MANAGED_IDENTITY_TOKEN_URL
-        = "http://169.254.169.254/metadata/identity/oauth2/token?api-version=2018-02-01";
+            = "http://169.254.169.254/metadata/identity/oauth2/token?api-version=2018-02-01";
 
     /**
      * Stores our logger.
@@ -71,16 +71,17 @@ class AuthClient extends DelegateRestClient {
      * Get an access token for a managed identity.
      *
      * @param resource the resource.
+     * @param identity the user-assigned identity (null if system-assigned)
      * @return the authorization token.
      */
-    public String getAccessToken(String resource) {
+    public String getAccessToken(String resource, String identity) {
         String result;
 
         if (System.getenv("WEBSITE_SITE_NAME") != null
             && !System.getenv("WEBSITE_SITE_NAME").isEmpty()) {
-            result = getAccessTokenOnAppService(resource);
+            result = getAccessTokenOnAppService(resource, identity);
         } else {
-            result = getAccessTokenOnOthers(resource);
+            result = getAccessTokenOnOthers(resource, identity);
         }
         return result;
     }
@@ -94,23 +95,24 @@ class AuthClient extends DelegateRestClient {
      * @param clientSecret the client secret.
      * @return the authorization token.
      */
-    public String getAccessToken(String resource, String tenantId,
-                                 String clientId, String clientSecret) {
-        LOGGER.entering("AuthClient", "getAccessToken", new Object[] {
-            resource, tenantId, clientId, clientSecret });
+    public String getAccessToken(String resource, String aadAuthenticationUrl,
+            String tenantId, String clientId, String clientSecret) {
+        
+        LOGGER.entering("AuthClient", "getAccessToken", new Object[]{
+            resource, tenantId, clientId, clientSecret});
         LOGGER.info("Getting access token using client ID / client secret");
         String result = null;
 
         StringBuilder oauth2Url = new StringBuilder();
-        oauth2Url.append(OAUTH2_TOKEN_BASE_URL)
-                 .append(tenantId)
-                 .append(OAUTH2_TOKEN_POSTFIX);
+        oauth2Url.append(aadAuthenticationUrl == null ? OAUTH2_TOKEN_BASE_URL : aadAuthenticationUrl)
+                .append(tenantId)
+                .append(OAUTH2_TOKEN_POSTFIX);
 
         StringBuilder requestBody = new StringBuilder();
         requestBody.append(GRANT_TYPE_FRAGMENT)
-                   .append(CLIENT_ID_FRAGMENT).append(clientId)
-                   .append(CLIENT_SECRET_FRAGMENT).append(clientSecret)
-                   .append(RESOURCE_FRAGMENT).append(resource);
+                .append(CLIENT_ID_FRAGMENT).append(clientId)
+                .append(CLIENT_SECRET_FRAGMENT).append(clientSecret)
+                .append(RESOURCE_FRAGMENT).append(resource);
 
         String body = post(oauth2Url.toString(), requestBody.toString(), "application/x-www-form-urlencoded");
         if (body != null) {
@@ -126,17 +128,24 @@ class AuthClient extends DelegateRestClient {
      * Get the access token on Azure App Service.
      *
      * @param resource the resource.
+     * @param identity the user-assigned identity (null if system-assigned).
      * @return the authorization token.
      */
-    private String getAccessTokenOnAppService(String resource) {
+    private String getAccessTokenOnAppService(String resource, String identity) {
         LOGGER.entering("AuthClient", "getAccessTokenOnAppService", resource);
         LOGGER.info("Getting access token using managed identity based on MSI_SECRET");
+        if (identity != null) {
+            LOGGER.log(INFO, "Using managed identity with object ID: {0}", identity);
+        }
         String result = null;
 
         StringBuilder url = new StringBuilder();
         url.append(System.getenv("MSI_ENDPOINT"))
            .append("?api-version=2017-09-01")
            .append(RESOURCE_FRAGMENT).append(resource);
+        if (identity != null) {
+            url.append("&objectid=").append(identity);
+        }
 
         HashMap<String, String> headers = new HashMap<>();
         headers.put("Metadata", "true");
@@ -156,16 +165,23 @@ class AuthClient extends DelegateRestClient {
      * Get the authorization token on everything else but Azure App Service.
      *
      * @param resource the resource.
+     * @param identity the user-assigned identity (null if system-assigned).
      * @return the authorization token.
      */
-    private String getAccessTokenOnOthers(String resource) {
+    private String getAccessTokenOnOthers(String resource, String identity) {
         LOGGER.entering("AuthClient", "getAccessTokenOnOthers", resource);
         LOGGER.info("Getting access token using managed identity");
+        if (identity != null) {
+            LOGGER.log(INFO, "Using managed identity with object ID: {0}", identity);
+        }
         String result = null;
 
         StringBuilder url = new StringBuilder();
         url.append(OAUTH2_MANAGED_IDENTITY_TOKEN_URL)
            .append(RESOURCE_FRAGMENT).append(resource);
+        if (identity != null) {
+            url.append("&object_id=").append(identity);
+        }
 
         HashMap<String, String> headers = new HashMap<>();
         headers.put("Metadata", "true");