+++ sdk/servicebus/azure-messaging-servicebus/src/test/java/com/azure/messaging/servicebus/FluxAutoLockRenewTest.java
@@ -56,9 +56,9 @@ public class FluxAutoLockRenewTest {
     private final ClientLogger logger = new ClientLogger(FluxAutoLockRenewTest.class);
 
     private final ServiceBusReceivedMessage receivedMessage = new ServiceBusReceivedMessage(BinaryData.fromString("Some Data"));
-    private final ServiceBusReceivedMessageContext message = new ServiceBusReceivedMessageContext(receivedMessage);
-    private final TestPublisher<ServiceBusReceivedMessageContext> messagesPublisher = TestPublisher.create();
-    private final Flux<? extends ServiceBusReceivedMessageContext> messageSource = messagesPublisher.flux();
+    private final ServiceBusMessageContext message = new ServiceBusMessageContext(receivedMessage);
+    private final TestPublisher<ServiceBusMessageContext> messagesPublisher = TestPublisher.create();
+    private final Flux<? extends ServiceBusMessageContext> messageSource = messagesPublisher.flux();
 
     private Function<String, Mono<OffsetDateTime>> renewalFunction;
 
@@ -105,7 +105,7 @@ public class FluxAutoLockRenewTest {
     void canCancel() {
         // Arrange
         final ServiceBusReceivedMessage receivedMessage2 = new ServiceBusReceivedMessage(BinaryData.fromString("data"));
-        final ServiceBusReceivedMessageContext message2 = new ServiceBusReceivedMessageContext(receivedMessage2);
+        final ServiceBusMessageContext message2 = new ServiceBusMessageContext(receivedMessage2);
         receivedMessage2.setLockToken(UUID.randomUUID());
         receivedMessage2.setLockedUntil(OffsetDateTime.now().plusSeconds(2));
 
@@ -225,7 +225,7 @@ public class FluxAutoLockRenewTest {
         final String expectedSessionId = "1";
         final FluxAutoLockRenew renewOperator = new FluxAutoLockRenew(messageSource,
             MAX_AUTO_LOCK_RENEW_DURATION, messageLockContainer, renewalFunction);
-        final ServiceBusReceivedMessageContext errorContext =  new ServiceBusReceivedMessageContext(expectedSessionId, new RuntimeException("fake error"));
+        final ServiceBusMessageContext errorContext =  new ServiceBusMessageContext(expectedSessionId, new RuntimeException("fake error"));
 
         when(messageLockContainer.addOrUpdate(eq(LOCK_TOKEN_STRING), any(OffsetDateTime.class), any(LockRenewalOperation.class)))
             .thenThrow(new RuntimeException("contained closed."));
@@ -333,13 +333,13 @@ public class FluxAutoLockRenewTest {
         receivedMessage2.setEnqueuedSequenceNumber(2);
         receivedMessage2.setLockToken(UUID.randomUUID());
         receivedMessage2.setLockedUntil(OffsetDateTime.now().plusSeconds(2));
-        final ServiceBusReceivedMessageContext message2 = new ServiceBusReceivedMessageContext(receivedMessage2);
+        final ServiceBusMessageContext message2 = new ServiceBusMessageContext(receivedMessage2);
 
         final ServiceBusReceivedMessage receivedMessage3 = new ServiceBusReceivedMessage(BinaryData.fromString("data"));
         receivedMessage3.setEnqueuedSequenceNumber(3);
         receivedMessage3.setLockToken(UUID.randomUUID());
         receivedMessage3.setLockedUntil(OffsetDateTime.now().plusSeconds(2));
-        final ServiceBusReceivedMessageContext message3 = new ServiceBusReceivedMessageContext(receivedMessage3);
+        final ServiceBusMessageContext message3 = new ServiceBusMessageContext(receivedMessage3);
 
 
         final FluxAutoLockRenew renewOperator = new FluxAutoLockRenew(messageSource,
@@ -372,12 +372,12 @@ public class FluxAutoLockRenewTest {
         final ServiceBusReceivedMessage receivedMessage2 = new ServiceBusReceivedMessage(BinaryData.fromString("data"));
         receivedMessage2.setLockToken(UUID.randomUUID());
         receivedMessage2.setLockedUntil(OffsetDateTime.now().plusSeconds(2));
-        final ServiceBusReceivedMessageContext message2 = new ServiceBusReceivedMessageContext(receivedMessage2);
+        final ServiceBusMessageContext message2 = new ServiceBusMessageContext(receivedMessage2);
 
         final ServiceBusReceivedMessage receivedMessage3 = new ServiceBusReceivedMessage(BinaryData.fromString("data"));
         receivedMessage3.setLockToken(UUID.randomUUID());
         receivedMessage3.setLockedUntil(OffsetDateTime.now().plusSeconds(2));
-        final ServiceBusReceivedMessageContext message3 = new ServiceBusReceivedMessageContext(receivedMessage3);
+        final ServiceBusMessageContext message3 = new ServiceBusMessageContext(receivedMessage3);
 
 
         final String expectedMappedValue = "New Expected Mapped Value";
@@ -407,13 +407,13 @@ public class FluxAutoLockRenewTest {
         final Long expectedEnqueuedSequenceNumber = 2L;
         final OffsetDateTime lockedUntil = OffsetDateTime.now().plusSeconds(1);
         final ServiceBusReceivedMessage receivedMessage2 = new ServiceBusReceivedMessage(BinaryData.fromString("data"));
-        final ServiceBusReceivedMessageContext message2 = new ServiceBusReceivedMessageContext(receivedMessage2);
+        final ServiceBusMessageContext message2 = new ServiceBusMessageContext(receivedMessage2);
         receivedMessage2.setLockToken(UUID.randomUUID());
         receivedMessage2.setLockedUntil(lockedUntil);
         receivedMessage2.setEnqueuedSequenceNumber(1);
 
         final ServiceBusReceivedMessage receivedMessage3 = new ServiceBusReceivedMessage(BinaryData.fromString("data"));
-        final ServiceBusReceivedMessageContext message3 = new ServiceBusReceivedMessageContext(receivedMessage3);
+        final ServiceBusMessageContext message3 = new ServiceBusMessageContext(receivedMessage3);
         receivedMessage2.setLockToken(UUID.randomUUID());
         receivedMessage2.setLockedUntil(lockedUntil);
         receivedMessage2.setEnqueuedSequenceNumber(expectedEnqueuedSequenceNumber);