+++ sdk/spring/azure-spring-boot-starter-keyvault-certificates/src/main/java/com/azure/spring/security/keyvault/certificates/starter/KeyVaultCertificatesEnvironmentPostProcessor.java
@@ -26,7 +26,7 @@ public class KeyVaultCertificatesEnvironmentPostProcessor implements Environment
 
     @Override
     public void postProcessEnvironment(ConfigurableEnvironment environment,
-                                       SpringApplication application) {
+            SpringApplication application) {
 
         Properties systemProperties = System.getProperties();
 
@@ -38,6 +38,11 @@ public class KeyVaultCertificatesEnvironmentPostProcessor implements Environment
             if (tenantId != null) {
                 systemProperties.put("azure.keyvault.tenantId", tenantId);
             }
+            
+            String aadAuthenticationUrl = environment.getProperty("azure.keyvault.aadAuthenticationUrl");
+            if (aadAuthenticationUrl != null) {
+                systemProperties.put("azure.keyvault.aadAuthenticationUrl", aadAuthenticationUrl);
+            }
 
             String clientId = environment.getProperty("azure.keyvault.clientId");
             if (clientId != null) {
@@ -47,6 +52,11 @@ public class KeyVaultCertificatesEnvironmentPostProcessor implements Environment
             String clientSecret = environment.getProperty("azure.keyvault.clientSecret");
             if (clientSecret != null) {
                 systemProperties.put("azure.keyvault.clientSecret", clientSecret);
+            }
+
+            String managedIdentity = environment.getProperty("azure.keyvault.managedIdentity");
+            if (managedIdentity != null) {
+                systemProperties.put("azure.keyvault.managedIdentity", managedIdentity);
             }
 
             String keyStoreType = environment.getProperty("server.ssl.key-store-type");
@@ -62,8 +72,8 @@ public class KeyVaultCertificatesEnvironmentPostProcessor implements Environment
                 } catch (ClassNotFoundException ex) {
                 }
 
-                PropertiesPropertySource propertySource =
-                        new PropertiesPropertySource("KeyStorePropertySource", properties);
+                PropertiesPropertySource propertySource
+                        = new PropertiesPropertySource("KeyStorePropertySource", properties);
                 sources.addFirst(propertySource);
             }
 
@@ -80,8 +90,8 @@ public class KeyVaultCertificatesEnvironmentPostProcessor implements Environment
                 } catch (ClassNotFoundException ex) {
                 }
 
-                PropertiesPropertySource propertySource = 
-                        new PropertiesPropertySource("TrustStorePropertySource", properties);
+                PropertiesPropertySource propertySource
+                        = new PropertiesPropertySource("TrustStorePropertySource", properties);
                 sources.addFirst(propertySource);
             }
 
@@ -90,8 +100,8 @@ public class KeyVaultCertificatesEnvironmentPostProcessor implements Environment
 
             String enabled = environment.getProperty("azure.keyvault.jca.overrideTrustManagerFactory");
             if (Boolean.parseBoolean(enabled)) {
-                KeyVaultTrustManagerFactoryProvider factoryProvider =
-                    new KeyVaultTrustManagerFactoryProvider();
+                KeyVaultTrustManagerFactoryProvider factoryProvider
+                        = new KeyVaultTrustManagerFactoryProvider();
                 Security.insertProviderAt(factoryProvider, 1);
             }