+++ sdk/spring/azure-spring-cloud-autoconfigure/src/main/java/com/azure/spring/cloud/autoconfigure/eventhub/AzureEventHubAutoConfiguration.java
@@ -74,12 +74,11 @@ public class AzureEventHubAutoConfiguration {
         } else {
             String connectionString = eventHubProperties.getConnectionString();
 
-            if (!StringUtils.hasText(connectionString)) {
-                throw new IllegalArgumentException("Event hubs connection string cannot be empty");
+            if (StringUtils.hasText(connectionString)) {
+                TelemetryCollector.getInstance()
+                                  .addProperty(EVENT_HUB, NAMESPACE, EventHubUtils.getNamespace(connectionString));
             }
 
-            TelemetryCollector.getInstance()
-                .addProperty(EVENT_HUB, NAMESPACE, EventHubUtils.getNamespace(connectionString));
             return new EventHubConnectionStringProvider(connectionString);
         }
     }